api/consultations/generate-suggestions機能を別ファイルに分離／接続するAzure MySQL DBをomukoroに変更
